@@ -18,13 +18,13 @@ namespace DotPulsar.Internal
 
     public static class Crc32C
     {
-        private const uint Generator = 0x82F63B78u;
+        private const uint _generator = 0x82F63B78u;
 
-        private static readonly uint[] Lookup;
+        private static readonly uint[] _lookup;
 
         static Crc32C()
         {
-            Lookup = new uint[16 * 256];
+            _lookup = new uint[16 * 256];
 
             for (uint i = 0; i < 256; i++)
             {
@@ -33,9 +33,9 @@ namespace DotPulsar.Internal
                 for (var j = 0; j < 16; j++)
                 {
                     for (var k = 0; k < 8; k++)
-                        entry = (entry & 1) == 1 ? Generator ^ (entry >> 1) : entry >> 1;
+                        entry = (entry & 1) == 1 ? _generator ^ (entry >> 1) : entry >> 1;
 
-                    Lookup[j * 256 + i] = entry;
+                    _lookup[j * 256 + i] = entry;
                 }
             }
         }
@@ -58,16 +58,16 @@ namespace DotPulsar.Internal
 
                     if (!readingBlock)
                     {
-                        checksum = Lookup[(byte) (checksum ^ currentByte)] ^ (checksum >> 8);
+                        checksum = _lookup[(byte) (checksum ^ currentByte)] ^ (checksum >> 8);
                         continue;
                     }
 
                     var offSetBase = offset * 256;
 
                     if (offset > 11)
-                        block[offset] = Lookup[offSetBase + ((byte) (checksum >> (8 * (15 - offset))) ^ currentByte)];
+                        block[offset] = _lookup[offSetBase + ((byte) (checksum >> (8 * (15 - offset))) ^ currentByte)];
                     else
-                        block[offset] = Lookup[offSetBase + currentByte];
+                        block[offset] = _lookup[offSetBase + currentByte];
 
                     --remaningBytes;
 
